Message backwards compatibility (#19)
* Set messages with the expected priority

The priority value is made up of the facility and severity values:
https://datatracker.ietf.org/doc/html/rfc5424#section-6.2.1

In moving from remote_syslog2 to go-loggregator/rfc5424 we dropped the
existing facility value. Set both facility and severity to match the
previous implementation.

* Create messages with the expected procid

In moving in #16 from papertrail/remote_syslog2 to
go-loggregator/rfc5424 we changed the procid emitted from "rs2" to
"cloudfoundry-blackbox". This was done to avoid misrepresenting
ourselves as remote_syslog2 after the change.

It turns out that the rs2 procid was actually added for blackbox in
4b9a249d.

There is a trade-off here between backwards compatibility and clarity,
we're going to favour backwards compatibility and keep the rs2 procid.

Note that it is important that the procid defined here matches the
procid in use by the syslog-release exclusion rule:
https://github.com/cloudfoundry/syslog-release/blob/b78ee21e566fbbd2f5a8253d7c3857ac51059dc1/jobs/syslog_forwarder/templates/syslog-release-file-exclusion.conf.erb

Co-authored-by: Andrew Crump <crumpan@vmware.com>